LayoutTests/fast/dom/Window/getMatchedCSSRules-with-pseudo-elements-expected.txt

Test for getMatchedCSSRules for Psuedo Elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pseudoRules[0].selectorText is selectorText
PASS pseudoRules[0].style.cssText.indexOf("1px") != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
